+++ documentation/Overview.asciidoc
@@ -59,7 +59,11 @@ and keep the page state intact. The following tutorials explain how to use the R
 * <<importing-dependencies/tutorial-ways-of-importing#,Ways of importing the dependencies>>
 
 == Theming Applications
-* <<theme/tutorial-built-in-themes#,Component themes>>
+* <<theme/theming-overview#,Theming Overview>>
+* <<theme/using-component-themes#,Using Component Themes>>
+* <<theme/application-theming-basics#,Application Theming Basics>>
+* <<theme/theming-crash-course#,Theming Crash Course>>
+* <<theme/integrating-component-theme#,Integrating Your Own Component Theme>>
 
 == Creating Components
 * <<creating-components/tutorial-component-basic#,Creating A Simple Component Using the Element API>>
