+++ documentation/src/main/html/ThemingCrashCourse.html
@@ -1,0 +1,91 @@
+tutorial::theme/theming-crash-course.asciidoc
+<custom-style>
+    <style>
+        /* Example global style */
+        html {
+            font-size: 12pt;
+        }
+    </style>
+</custom-style>
+<link rel="import" href="../bower_components/polymer/polymer-element.html">
+
+<dom-module id="my-view">
+    <template>
+        <style>
+            /* Example scoped styles, applied to <my-view> template contents only */
+
+            :host {
+                /* Styles for the `<my-view>` hosting element */
+                display: block;
+            }
+
+            .my-view-title {
+                font-weight: bold;
+                border-bottom: 1px solid gray;
+            }
+        </style>
+
+        <div class="my-view-title">My view title</div>
+    </template>
+</dom-module>
+<script>
+    class MyView extends Polymer.Element {
+        static get is() { return 'my-view'; }
+    }
+
+    customElements.define(MyView.is, MyView);
+</script>
+<custom-style>
+    <style>
+        html {
+            /* Example global custom CSS property definition */
+            --my-theme-color: brown;
+        }
+    </style>
+</custom-style>
+<!-- ... -->
+
+<dom-module id="my-view">
+    <template>
+        <style>
+            /* ... */
+
+            .my-view-title {
+                /* Example referencing custom CSS property */
+                color: var(--my-theme-color);
+            }
+        </style>
+    </template>
+
+    <!-- ... -->
+</dom-module>
+
+<dom-module id="shared-styles">
+    <template>
+        <style>
+            /* Example style module */
+            .my-outline-style {
+                outline: 1px solid green;
+            }
+        </style>
+    </template>
+</dom-module>
+
+<!-- ... -->
+<link rel="import" href="../styles/shared-styles.html">
+
+<dom-module id="my-view">
+    <template>
+        <style include="shared-styles">
+            /*  */
+        </style>
+    </template>
+
+    <!-- ... -->
+</dom-module>
+
+<link rel="import" href="./shared-styles.html">
+
+<custom-style>
+    <style include="shared-styles"></style>
+</custom-style>
